test/ui-pytest/requirements.txt
pytest==7.4.0
psutil==5.9.5
pillow==10.0.0
opencv-python-headless==4.8.0.74
numpy~=1.25.1
pytesseract==0.3.10
atomacos==3.3.0; platform_system == "Darwin"
allure-pytest==2.13.2
testrail-api==1.12.0
